feat: add admin funcionality

+++ server/src/routes/admin.ts
@@ -10,6 +10,7 @@ import { eq } from 'drizzle-orm';
 import { is_admin } from '../players';
 import { hide_route } from './common';
 
+const MUNOST_STEAM_ID = '76561198377195635';
 const app = new Hono<{ Variables: Variables }>();
 
 const BooleanInput = z.object({
@@ -24,6 +25,10 @@ app.post(
   async (c) => {
     const maintenance = c.req.valid('json').new_value;
     server_state.maintenance = maintenance;
+
+    return c.json({
+      data: `Maintenace was toggled, new value: ${maintenance}`,
+    });
   }
 );
 
@@ -34,8 +39,13 @@ app.post(
   zValidator('json', BooleanInput),
   async (c) => {
     const steam_id = c.req.param('steam_id');
+
+    if (steam_id === MUNOST_STEAM_ID) {
+      return c.json({ error: 'Nice try' }, 400);
+    }
+
     const new_value = c.req.valid('json').new_value;
-    const admin = is_admin(steam_id);
+    const admin = await is_admin(steam_id);
 
     if (admin && !new_value) {
       await db.delete(admins).where(eq(admins.steam_id, steam_id));
@@ -44,15 +54,19 @@ app.post(
         steam_id,
       });
     }
+
+    return c.json({
+      data: `Player's admin was toggled, new value: ${!admin}`,
+    });
   }
 );
 
 app.get('/player/:steam_id', hide_route(), admin_auth, async (c) => {
   const steam_id = c.req.param('steam_id');
-  const admin = is_admin(steam_id);
+  const admin = await is_admin(steam_id);
 
   return c.json({
-    admin,
+    data: admin,
   });
 });
